requirements.txt
adium-theme-ubuntu (0.3.4)
appdirs (1.4.3)
backports.shutil-get-terminal-size (1.0.0)
backports.weakref (1.0rc1)
biosppy (0.4.0)
bleach (1.5.0)
ccsm (0.9.12.2)
certifi (2017.4.17)
chardet (3.0.4)
compizconfig-python (0.9.12.2)
configparser (3.5.0)
cycler (0.10.0)
Cython (0.25.2)
dask (0.14.0)
decorator (4.0.10)
easydict (1.6)
entrypoints (0.2.2)
enum34 (1.1.6)
funcsigs (1.0.2)
functools32 (3.2.3.post2)
h5py (2.6.0)
html5lib (0.9999999)
idna (2.5)
ipykernel (4.5.2)
ipython (5.1.0)
ipython-genutils (0.1.0)
ipywidgets (6.0.0)
Jinja2 (2.8)
jsonschema (2.6.0)
jupyter (1.0.0)
jupyter-client (5.0.0)
jupyter-console (5.1.0)
jupyter-core (4.3.0)
Keras (2.0.5)
Mako (1.0.6)
Markdown (2.2.0)
MarkupSafe (0.23)
matplotlib (1.5.3)
mistune (0.7.3)
mock (2.0.0)
nbconvert (5.1.1)
nbformat (4.3.0)
networkx (1.11)
nose (1.3.7)
notebook (4.4.1)
numpy (1.13.0)
olefile (0.44)
opencv-python (3.2.0.6)
packaging (16.8)
pandas (0.20.2)
pandocfilters (1.4.1)
pathlib2 (2.1.0)
pbr (3.0.1)
pexpect (4.2.1)
pickleshare (0.7.4)
Pillow (3.3.1)
pip (9.0.1)
prompt-toolkit (1.0.7)
protobuf (3.0.0b2)
ptyprocess (0.5.1)
pycurl (7.43.0)
pydot (1.1.0)
pygame (1.9.3)
Pygments (2.1.3)
pygpu (0.6.5)
PyMySQL (0.7.2)
pyparsing (2.1.9)
Pyste (0.9.10)
python-dateutil (2.5.3)
pytz (2016.6.1)
PyYAML (3.12)
pyzmq (15.2.0)
qtconsole (4.2.1)
requests (2.17.3)
scandir (1.5)
scikit-image (0.12.3)
scikit-learn (0.17.1)
scipy (0.19.0)
seaborn (0.7.1)
setuptools (36.0.1)
shortuuid (0.5.0)
simplegeneric (0.8.1)
six (1.10.0)
sklearn (0.0)
subprocess32 (3.2.7)
tensorflow-gpu (1.2.0)
terminado (0.6)
testpath (0.3)
Theano (0.9.0)
toolz (0.8.2)
tornado (4.2.1)
traitlets (4.3.1)
unity-lens-photos (1.0)
update (0.4.4)
urllib3 (1.21.1)
virtualenv (15.1.0)
wcwidth (0.1.7)
webencodings (0.5)
Werkzeug (0.12.2)
wfdb (1.1.2)
wheel (0.29.0)
widgetsnbextension (2.0.0)
